server.py: Added ability to use a bind user
Previous configuration only supported anonymous binds which were
problematic with a  Windows server which filters out any admin
user from the search and stores 'cn' as the user's real name. This
patch should be compatible with LDAP since both schemas use the
sAMAccountName for the username.  How this works is first using
the bind user to search for the intended user, and then attempt
to re-connect using the user's DN (from the previous search)
and the provided password.

Signed-off-by: Brian T. Goodwin <btgoodwin@tenacity.geontech.local>